Support AAudio based input recording to enable full duplex AAudio support
* Initial implementation of AAudio based input recording
* Verified duplex operation on Pixel 3
* Made unit test mode working with openSL mock, as before
* Updated example activity
* Fix handlers
* Buffer size and temporary buffers now update during runtime buffer size updates. Use preferred floating point format for AAUdio, use memory cloning for faster writes.
* README and inline doc copy updates
* and again
* Fall back to OpenSL default

+++ src/main/cpp/events/basesynthevent.cpp
@@ -189,7 +189,7 @@ void BaseSynthEvent::calculateBuffers()
 
     if ( isSequenced )
     {
-        setEventStart( position * ( int ) AudioEngine::samples_per_step );
+        setEventStart( position * AudioEngine::samples_per_step );
         setEventLength(( int )( length * AudioEngine::samples_per_step ));
         setEventEnd( _eventStart + _eventLength );
     }
